refactor(sampledecryptdec): fix build error on Android ndk 23
(cherry picked from commit e6a958ad974632cd8b67d720f41fb1e3073d6a18)

+++ framework/demuxer/sample_decrypt/sampleDecryptDec.c
@@ -158,7 +158,7 @@ static int sampleDecrypt_read_close(AVFormatContext *s)
 #define VD AV_OPT_FLAG_VIDEO_PARAM | AV_OPT_FLAG_DECODING_PARAM
 static const AVOption options[] = {
     {
-        "Decryptor", "Decryptor address", OFFSET(decryptor), AV_OPT_TYPE_UINT64, {.i64 = 0}, INT64_MIN, INT64_MAX, VD
+        "Decryptor", "Decryptor address", OFFSET(decryptor), AV_OPT_TYPE_UINT64, {.i64 = 0}, INT64_MIN, (double)INT64_MAX, VD
     },
 
     {NULL}
